+++ source/SU3_ops.c
@@ -6,7 +6,7 @@
 #include "math_ops.h"  //	Math operations
 #include "lattice.h"
 
-void SU3_print_matrix(const double complex *u, const char *name) {
+void SU3_print_matrix(const float complex *u, const char *name) {
     // Prints the matrix on screen
 
     printf("\n\n %s \n", name);
@@ -29,7 +29,7 @@ void SU3_print_matrix(const double complex *u, const char *name) {
     getchar();
 }
 
-void SU3_copy(const double complex *u, double complex *u_copy) {
+void SU3_copy(const float complex *u, float complex *u_copy) {
     // Copies u to u_copy
 
     for (unsigned short a = 0; a < 3; a++) {
@@ -39,7 +39,7 @@ void SU3_copy(const double complex *u, double complex *u_copy) {
     }
 }
 
-void SU3_set_to_null(double complex *u) {
+void SU3_set_to_null(float complex *u) {
     // Sets u to be the null matrix in SU(3)
 
     for (unsigned short a = 0; a < 3; a++) {
@@ -49,7 +49,7 @@ void SU3_set_to_null(double complex *u) {
     }
 }
 
-void SU3_set_to_identity(double complex *u) {
+void SU3_set_to_identity(float complex *u) {
     // Sets u to be the identity matrix in SU(3)
 
     for (unsigned short a = 0; a < 3; a++) {
@@ -61,7 +61,7 @@ void SU3_set_to_identity(double complex *u) {
     }
 }
 
-void SU3_accumulate(const double complex *u, double complex *acc) {
+void SU3_accumulate(const float complex *u, float complex *acc) {
     // Accumulates the value of u into acc
 
     for (unsigned short a = 0; a < 3; a++) {
@@ -73,8 +73,8 @@ void SU3_accumulate(const double complex *u, double complex *acc) {
     }
 }
 
-void SU3_subtraction(const double complex *u, const double complex *v,
-                     double complex *u_minus_v) {
+void SU3_subtraction(const float complex *u, const float complex *v,
+                     float complex *u_minus_v) {
     //  Calculates the difference between matrix u and matrix v
     //  and returns result in u_minus_v
 
@@ -87,14 +87,14 @@ void SU3_subtraction(const double complex *u, const double complex *v,
     }
 }
 
-double complex SU3_trace(const double complex *u) {
+float complex SU3_trace(const float complex *u) {
     //	Calculates the trace of the matrix u
     //  and returns result (complex) in tr
 
     return u[0 * 3 + 0] + u[1 * 3 + 1] + u[2 * 3 + 2];
 }
 
-double complex SU3_determinant(const double complex *u) {
+float complex SU3_determinant(const float complex *u) {
     //  Calculates the determinant of the matrix u
     //  and returns result, a complex number, in det.
 
@@ -104,7 +104,7 @@ double complex SU3_determinant(const double complex *u) {
     return u[0 * 3 + 0] * (u[1 * 3 + 1] * u[2 * 3 + 2] - u[1 * 3 + 2] * u[2 * 3 + 1]) + u[0 * 3 + 1] * (u[1 * 3 + 2] * u[2 * 3 + 0] - u[1 * 3 + 0] * u[2 * 3 + 2]) + u[0 * 3 + 2] * (u[1 * 3 + 0] * u[2 * 3 + 1] - u[1 * 3 + 1] * u[2 * 3 + 0]);
 }
 
-void SU3_hermitean_conjugate(const double complex *u, double complex *u_dagger) {
+void SU3_hermitean_conjugate(const float complex *u, float complex *u_dagger) {
     // Calculates the hermitean conjugate to u
     // and returns result in u_dagger.
 
@@ -124,8 +124,8 @@ void SU3_hermitean_conjugate(const double complex *u, double complex *u_dagger) 
     }
 }
 
-void SU3_multiplication_by_scalar(const double complex alpha, const double complex *u,
-                                  double complex *alpha_times_u) {
+void SU3_multiplication_by_scalar(const float complex alpha, const float complex *u,
+                                  float complex *alpha_times_u) {
     //  Calculates multiplication of SU(3) matrix u by scalar alpha
     //  and returns result in alphatimesu.
 
@@ -139,7 +139,7 @@ void SU3_multiplication_by_scalar(const double complex alpha, const double compl
     }
 }
 
-void SU3_substitution_multiplication_by_scalar(const double complex alpha, double complex *u) {
+void SU3_substitution_multiplication_by_scalar(const float complex alpha, float complex *u) {
     //  Calculates multiplication of SU(3) matrix u by scalar alpha
     //  and returns result in u.
 
@@ -153,7 +153,7 @@ void SU3_substitution_multiplication_by_scalar(const double complex alpha, doubl
     }
 }
 
-void SU3_product(const double complex *u, const double complex *v, double complex *uv) {
+void SU3_product(const float complex *u, const float complex *v, float complex *uv) {
     // Calculates product of 2 SU(3) matrices u e v
     // and returns result in uv.
 
@@ -169,8 +169,8 @@ void SU3_product(const double complex *u, const double complex *v, double comple
     }
 }
 
-void SU3_product_three(const double complex *u, const double complex *v, const double complex *w,
-                       double complex *uvw) {
+void SU3_product_three(const float complex *u, const float complex *v, const float complex *w,
+                       float complex *uvw) {
     //  Calculates product of 3 SU(3) matrices u, v and w
     //  and returns result in uvw.
 
@@ -188,31 +188,31 @@ void SU3_product_three(const double complex *u, const double complex *v, const d
     }
 }
 
-void SU3_accumulate_left_product(const double complex *g, double complex *acc_prod) {
+void SU3_accumulate_left_product(const float complex *g, float complex *acc_prod) {
     //	Calculates matrix product between g and acc_prod and accumulates result in acc_prod
-    double complex aux_prod[3 * 3];
+    float complex aux_prod[3 * 3];
  
     SU3_copy(acc_prod, aux_prod);
     SU3_product(g, aux_prod, acc_prod);
 
 }
 
-void SU3_accumulate_right_product(double complex *acc_prod, const double complex *g) {
+void SU3_accumulate_right_product(float complex *acc_prod, const float complex *g) {
     //	Calculates matrix product between acc_prod and g and accumulates result in acc_prod
 
-    double complex aux_prod[3 * 3];
+    float complex aux_prod[3 * 3];
 
     SU3_copy(acc_prod, aux_prod);
     SU3_product(aux_prod, g, acc_prod);
 
 }
 
-void SU3_projection(double complex *x) {
+void SU3_projection(float complex *x) {
     //	Projects matrix u to the group SU(3) returning SU(3) matrix x at the end.
     //	Follows method found in Gattringer around Eq. 4.27.
     //  More explanation below.
 
-    double sum_absvalue = 0.0;  //  To calculate first two lines of
+    float sum_absvalue = 0.0;  //  To calculate first two lines of
                                 //  projected matrix.
 
     for (unsigned short b = 0; b < 3; b++) {
@@ -221,15 +221,15 @@ void SU3_projection(double complex *x) {
         //  to be used for normalization below.
     }
 
-    double complex x_SU3[3 * 3];
+    float complex x_SU3[3 * 3];
 
     //  Used in the Gram-Schmidt
-    double complex v_unewconj = 0.0;  //  method to calculate
+    float complex v_unewconj = 0.0;  //  method to calculate
                                       //  second line of projected
                                       //  matrix.
     
-    double complex u_new_conj[3];  //  To calculate last line of
-    double complex v_new_conj[3];  //  projected matrix.
+    float complex u_new_conj[3];  //  To calculate last line of
+    float complex v_new_conj[3];  //  projected matrix.
 
     for (unsigned short b = 0; b < 3; b++) {
         x_SU3[0 * 3 + b] = x[0 * 3 + b] / sqrt(sum_absvalue);
@@ -244,7 +244,7 @@ void SU3_projection(double complex *x) {
         //  to be used below for the Gram-Schmidt method.
     }
 
-    double complex v_prime[3];
+    float complex v_prime[3];
     //	Used in the Gram-Schmidt  method to calculate
     //  second line of projected matrix.
 
@@ -288,7 +288,7 @@ void SU3_projection(double complex *x) {
 
 }
 
-void SU3_decompose_algebra(const double complex *a, double *a_components) {
+void SU3_decompose_algebra(const float complex *a, float *a_components) {
     //  Decomposes A in the components of the alfebra of SU(3)
     //  using the Gell-Mann matrices a basis and following the conventions
     //  of Gattringer. The components are then returned in a_components
